requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
colorama==0.4.6
greenlet==2.0.2
iniconfig==2.0.0
loguru==0.7.0
packaging==23.1
playwright==1.37.0
pluggy==1.2.0
pyee==9.0.4
pytest==7.4.0
PyYAML==6.0.1
typing_extensions==4.7.1
win32-setctime==1.1.0
